Move VirtualColumn, VirtualColumns back to io.druid.segment package.
