+++ processing/src/main/java/io/druid/segment/CursorFactory.java
@@ -22,7 +22,6 @@ package io.druid.segment;
 import io.druid.granularity.QueryGranularity;
 import io.druid.java.util.common.guava.Sequence;
 import io.druid.query.filter.Filter;
-import io.druid.segment.virtual.VirtualColumns;
 import org.joda.time.Interval;
 
 /**
